refactor: restructure backend, clean unused files, and finalize core chatbot logic
