--- notebooks/requirements_working.txt
@@ -1,171 +1,0 @@
-aiohappyeyeballs==2.6.1
-aiohttp==3.11.18
-aiosignal==1.3.2
-altair==5.5.0
-annotated-types==0.7.0
-anyio==4.9.0
-appnope==0.1.4
-argon2-cffi==23.1.0
-argon2-cffi-bindings==21.2.0
-arrow==1.3.0
-asttokens==3.0.0
-async-lru==2.0.5
-attrs==25.3.0
-babel==2.17.0
-beautifulsoup4==4.13.4
-black==24.3.0
-bleach==6.2.0
-blinker==1.9.0
-cachetools==5.5.2
-certifi==2025.4.26
-cffi==1.17.1
-charset-normalizer==3.4.1
-click==8.1.8
-comm==0.2.2
-dataclasses-json==0.6.7
-debugpy==1.8.14
-decorator==5.2.1
-defusedxml==0.7.1
-distro==1.9.0
-executing==2.2.0
-faiss-cpu==1.9.0
-fastjsonschema==2.21.1
-filelock==3.18.0
-fqdn==1.5.1
-frozenlist==1.6.0
-fsspec==2025.3.2
-gitdb==4.0.12
-GitPython==3.1.44
-groq==0.23.1
-h11==0.16.0
-httpcore==1.0.9
-httpx==0.28.1
-httpx-sse==0.4.0
-huggingface-hub==0.30.2
-idna==3.10
-ipykernel==6.29.5
-ipython==9.2.0
-ipython_pygments_lexers==1.1.1
-ipywidgets==8.0.6
-isoduration==20.11.0
-jedi==0.19.2
-Jinja2==3.1.6
-joblib==1.4.2
-json5==0.12.0
-jsonpatch==1.33
-jsonpointer==3.0.0
-jsonschema==4.23.0
-jsonschema-specifications==2025.4.1
-jupyter-black==0.4.0
-jupyter-events==0.12.0
-jupyter-lsp==2.2.5
-jupyter_client==8.6.3
-jupyter_core==5.7.2
-jupyter_server==2.15.0
-jupyter_server_terminals==0.5.3
-jupyterlab==4.2.5
-jupyterlab_pygments==0.3.0
-jupyterlab_server==2.27.3
-jupyterlab_widgets==3.0.14
-langchain==0.3.3
-langchain-community==0.3.2
-langchain-core==0.3.56
-langchain-groq==0.2.0
-langchain-huggingface==0.1.0
-langchain-mistralai==0.2.10
-langchain-text-splitters==0.3.8
-langchainhub==0.1.21
-langsmith==0.1.147
-MarkupSafe==3.0.2
-marshmallow==3.26.1
-matplotlib-inline==0.1.7
-mistune==3.1.3
-mpmath==1.3.0
-multidict==6.4.3
-mypy_extensions==1.1.0
-narwhals==1.37.1
-nbclient==0.10.2
-nbconvert==7.16.6
-nbformat==5.10.4
-nest-asyncio==1.6.0
-networkx==3.4.2
-notebook_shim==0.2.4
-numpy==1.26.4
-orjson==3.10.16
-overrides==7.7.0
-packaging==24.2
-pandas==2.2.1
-pandocfilters==1.5.1
-parso==0.8.4
-pathspec==0.12.1
-pexpect==4.9.0
-pillow==11.2.1
-platformdirs==4.3.7
-prometheus_client==0.21.1
-prompt_toolkit==3.0.51
-propcache==0.3.1
-protobuf==6.30.2
-psutil==7.0.0
-ptyprocess==0.7.0
-pure_eval==0.2.3
-pyarrow==20.0.0
-pycparser==2.22
-pydantic==2.11.3
-pydantic-settings==2.9.1
-pydantic_core==2.33.1
-pydeck==0.9.1
-Pygments==2.19.1
-python-dateutil==2.9.0.post0
-python-dotenv==1.0.1
-python-json-logger==3.3.0
-pytz==2025.2
-PyYAML==6.0.2
-pyzmq==26.4.0
-referencing==0.36.2
-regex==2024.11.6
-requests==2.32.3
-requests-toolbelt==1.0.0
-rfc3339-validator==0.1.4
-rfc3986-validator==0.1.1
-rpds-py==0.24.0
-safetensors==0.5.3
-scikit-learn==1.6.1
-scipy==1.15.2
-Send2Trash==1.8.3
-sentence-transformers==4.1.0
-sentencepiece==0.2.0
-six==1.17.0
-smmap==5.0.2
-sniffio==1.3.1
-soupsieve==2.7
-SQLAlchemy==2.0.40
-stack-data==0.6.3
-streamlit==1.45.0
-sympy==1.14.0
-tenacity==8.5.0
-terminado==0.18.1
-testbook==0.4.2
-threadpoolctl==3.6.0
-tinycss2==1.4.0
-tokenize_rt==6.1.0
-tokenizers==0.15.2
-toml==0.10.2
-torch==2.0.0
-tornado==6.4.2
-tqdm==4.67.1
-traitlets==5.14.3
-transformers==4.38.2
-types-python-dateutil==2.9.0.20241206
-types-requests==2.32.0.20250328
-typing-inspect==0.9.0
-typing-inspection==0.4.0
-typing_extensions==4.13.2
-tzdata==2025.2
-uri-template==1.3.0
-urllib3==2.4.0
-wcwidth==0.2.13
-webcolors==24.11.1
-webencodings==0.5.1
-websocket-client==1.8.0
-widgetsnbextension==4.0.14
-yarl==1.20.0
