+++ requirements.txt
@@ -24,3 +24,5 @@ langchain-community==0.3.2
 langchain-huggingface==0.1.0
 langgraph==0.4.2
 faiss-cpu==1.9.0
+uvicorn==0.34.2
+fastapi==0.115.12
